Update SSHW1.c
Finished the toInstruction() function that returns an array of instruction structs but I don't have a way of making sure that it works yet. I'll probably have to write a function to print the array in case we have a problem and need to debug down the line. Please let me know what you think.