Update SSHW1.c
I solved 90% of our file read problems (the last part of the problem is on line 190) and I figure the next item on the to do list is figuring out the final structure of fetch and execute because we need to call fetch from inside of the while loop on line 86 if we want it to work. We're making progress!